cleanup and add arch diagram
