+++ lib/crossplane-argocd-gitops/multi-cluster-pipeline.ts
@@ -1,7 +1,9 @@
 import { Construct } from "constructs";
 import * as blueprints from '@aws-quickstart/eks-blueprints';
 import {K8S_VERSIONS_DEV, MultiClusterOptions} from "./multi-cluster-options";
+import {CapacityType, KubernetesVersion} from "aws-cdk-lib/aws-eks";
 import {NodegroupAmiType} from "aws-cdk-lib/aws-eks";
+import * as eks from "aws-cdk-lib/aws-eks";
 import * as ec2 from "aws-cdk-lib/aws-ec2";
 import ManagementClusterBuilder from "./management-cluster-builder";
 import {ProviderMgmtRoleTeam} from "./custom-addons/mgmt-role-teams";
@@ -10,8 +12,39 @@ import {IRole} from "aws-cdk-lib/aws-iam";
 import * as iam from 'aws-cdk-lib/aws-iam';
 import {ManagedNodeGroup} from "@aws-quickstart/eks-blueprints/dist/cluster-providers/types";
 
+const account = process.env.CDK_DEFAULT_ACCOUNT ?? "";
+//const region = process.env.CDK_DEFAULT_REGION ?? "us-east-1";
+const region = process.env.CDK_DEFAULT_REGION!;
+const minSize  =  parseInt(process.env.NODEGROUP_MIN ?? "1");
+const maxSize  =  parseInt(process.env.NODEGROUP_MAX ?? "3");
+const desiredSize  =  parseInt(process.env.NODEGROUP_DESIRED ?? "1");
+const gitHubSecret = process.env.GITHUB_SECRET ?? "cdk_blueprints_github_secret";
+
+const props : MultiClusterOptions = {
+    account,
+    region,
+    minSize,
+    maxSize,
+    desiredSize,
+    gitHubSecret,
+    nodeGroupCapacityType: CapacityType.ON_DEMAND,
+    k8sVersions: K8S_VERSIONS_DEV // K8S_VERSIONS_PROD for full deploy
+};
+
+
+const mngProps: blueprints.MngClusterProviderProps = {
+    version: KubernetesVersion.V1_28,
+    instanceTypes: [ec2.InstanceType.of(ec2.InstanceClass.M5, ec2.InstanceSize.XLARGE2)],
+    amiType: eks.NodegroupAmiType.AL2_X86_64,
+    desiredSize: 2,
+    maxSize: 3,
+};
+
+console.info("Running CDK with id: crossplane-argocd-gitops" );
+console.info("Running CDK with: " + JSON.stringify(props));
+
 export default class MultiClusterPipelineConstruct {
-    async buildAsync(scope: Construct, id: string, props: MultiClusterOptions, mngProps: blueprints.MngClusterProviderProps) {
+    async buildAsync(scope: Construct, id: string) {
         const k8sVersions = props.k8sVersions ?? K8S_VERSIONS_DEV;
         const region :string = props.region;
         const account : string = props.account;
